refactor(adf): dismantle monolithic types.ts into domain-owned modules
Split types.ts (258 LOC) into 6 focused domain files under types/:
- ast.ts: AdfDocument, AdfSection, AdfContent, AdfMapEntry, AdfMetricEntry
- decorations.ts: STANDARD_DECORATIONS, CANONICAL_KEY_ORDER
- patch.ts: PatchOperation union and 7 op interfaces
- manifest.ts: Manifest, ManifestModule, SyncEntry, CadenceEntry, MetricSource
- bundle.ts: BundleResult
- validation.ts: ConstraintStatus, ConstraintResult, WeightSummary, EvidenceResult, AdfLockfile, AdfSyncStatus

Original types.ts replaced with barrel re-export for backward compatibility.
Zero import changes needed across all consumers.

Closes #10
