+++ packages/adf/src/types/validation.ts
@@ -1,0 +1,46 @@
+/**
+ * ADF Validation — constraint results, weight summaries, evidence, sync/lockfile types.
+ */
+
+export type ConstraintStatus = 'pass' | 'fail' | 'warn';
+
+export interface ConstraintResult {
+  section: string;
+  metric: string;
+  value: number;
+  ceiling: number;
+  unit: string;
+  status: ConstraintStatus;
+  message: string;
+  source: 'metric' | 'context';
+}
+
+export interface WeightSummary {
+  loadBearing: number;
+  advisory: number;
+  unweighted: number;
+  total: number;
+}
+
+export interface EvidenceResult {
+  constraints: ConstraintResult[];
+  weightSummary: WeightSummary;
+  allPassing: boolean;
+  failCount: number;
+  warnCount: number;
+}
+
+/**
+ * .adf.lock file format: flat JSON map of source filename -> sha256 prefix (16 hex chars).
+ * Generated by `charter adf sync --write`, checked by `charter adf sync --check`.
+ */
+export interface AdfLockfile {
+  [sourceFile: string]: string;
+}
+
+export interface AdfSyncStatus {
+  source: string;
+  sourceHash: string;
+  lockedHash: string | null;
+  inSync: boolean;
+}
